+++ CaseStudy6/src/main/java/com/example/casestudy6/repository/IRoleRepo.java
@@ -3,6 +3,8 @@ package com.example.casestudy6.repository;
 import com.example.casestudy6.model.Role;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface IRoleRepo extends PagingAndSortingRepository<Role,Long> {
+import javax.persistence.criteria.CriteriaBuilder;
+
+public interface IRoleRepo extends PagingAndSortingRepository<Role, Integer> {
     Role findByName(String roleName);
 }
